+++ axelor-contract/src/main/java/com/axelor/apps/contract/service/ContractServiceImpl.java
@@ -389,8 +389,7 @@ public class ContractServiceImpl extends ContractRepository implements ContractS
   @Override
   @Transactional(rollbackOn = {AxelorException.class, RuntimeException.class})
   public Invoice invoicingContract(Contract contract) throws AxelorException {
-    InvoiceGenerator invoiceGenerator = new InvoiceGeneratorContract(contract);
-    Invoice invoice = invoiceGenerator.generate();
+    Invoice invoice = generateInvoice(contract);
     InvoiceRepository invoiceRepository = Beans.get(InvoiceRepository.class);
     invoiceRepository.save(invoice);
 
@@ -471,6 +470,13 @@ public class ContractServiceImpl extends ContractRepository implements ContractS
     increaseInvoiceDates(contract);
 
     return invoiceRepository.save(invoice);
+  }
+
+  public Invoice generateInvoice(Contract contract) throws AxelorException {
+    InvoiceGenerator invoiceGenerator = new InvoiceGeneratorContract(contract);
+    Invoice invoice = invoiceGenerator.generate();
+
+    return invoice;
   }
 
   @Override