+++ CHANGELOG.md
@@ -122,6 +122,7 @@
 - PRICE LIST : hide O2M historized price lists in pop-up view
 - MARKETING : Precise domain for model in message template
 - Change titles for productFamily.
+- CONTRACT : Set project on generated invoices
 
 ## Bug Fixes
 - TeamTask : Resolve NPE on save of new team task.